Some refactoring of estimation code
The big new function here is a single "prepare model" function that returns everything of use for a particular specification: the estimation agents, empirical moments, moment weights; and functions to evaluate the objective function, generate simulated moments, or plot simulated vs empirical moments. These aren't used in the estimation right now, but is useful for experimenting with the models and could be used to make a non-estimagic estimator.

I also added moment weights that depend on the number of SCF observations (sqrt), which significantly downweights later ages.

There are three sets of estimations here (which should have better labels). The ones with no letter are equivalent to the specs we sent to TRP, but with simulation corrected and the new weights. The version with X removes the bespoke PermGroFac=0.94 at retirement, leaving it at its default value. The version with Z does that *and* allows DiscFac to float.

I now realize that the code is pointing to the non-local estimark when importing parameters, so my (small) changes to parameters.py are actually in *that* repository. I will try to fix.