More paper editing, ran a lot more specs
Paper has been mostly edited, but we still need to nail down a specification. I've run a *lot* of them, and none are really satisfying. The one labeled W seems like it might be the best compromise on simplicity of specification and results.

I still need to add the expense shocks in retirement from Fulford and Low, and maybe run the "subjective expectations" analysis.